requirements.txt
  fastapi
  uvicorn
  sqlalchemy
  pymysql
  alembic
  python-dotenv
  pydantic
  pydantic[email]
  bcrypt
  python-jose[cryptography]
  pytest
  redis